Rename tests again, make folders for them
