+++ tests/connection/noConnection.spec.ts
@@ -1,5 +1,5 @@
 import {test, expect} from "@playwright/test"
-import { sendMessage } from "./testHelpers.spec";
+import { sendMessage } from ".././testHelpers.spec";
 
 test("adding a message works", async ({ page }) => {
   // TODO:
