+++ tests/movingChannels/searchingForChannel.spec.ts
